backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.1
transformers==4.38.2
sentencepiece==0.1.99
webcolors==1.13
colorthief==0.2.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.1.0
textstat==0.7.3
scikit-image==0.22.0
selenium==4.17.2
playwright==1.41.0
axe-core-python==4.9.0
pa11y==6.0.0
openai==1.12.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.1
PyGithub==1.59.1
jwt==1.3.1
diff-match-patch==20230430

